# week-05/report_assets/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "3D Printer Production Simulator",
        "description": "3D Printer Production Simulator API — manage inventory, orders, and production cycles.",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/config": {
            "get": {
                "tags": [
                    "configuration"
                ],
                "summary": "Get Config",
                "description": "Get full production configuration (models, BOMs, suppliers).",
                "operationId": "get_config_api_config_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConfigResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/config/models": {
            "get": {
                "tags": [
                    "configuration"
                ],
                "summary": "List Models",
                "description": "List all product models.",
                "operationId": "list_models_api_config_models_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/ProductModelSchema"
                                    },
                                    "type": "array",
                                    "title": "Response List Models Api Config Models Get"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/config/models/{model_id}": {
            "get": {
                "tags": [
                    "configuration"
                ],
                "summary": "Get Model",
                "description": "Get a specific product model with its full BOM.",
                "operationId": "get_model_api_config_models__model_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "model_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Model Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductModelSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/config/suppliers": {
            "get": {
                "tags": [
                    "configuration"
                ],
                "summary": "List Suppliers",
                "description": "List all active suppliers with their product catalogs.",
                "operationId": "list_suppliers_api_config_suppliers_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/SupplierSchema"
                                    },
                                    "type": "array",
                                    "title": "Response List Suppliers Api Config Suppliers Get"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "configuration"
                ],
                "summary": "Add Supplier",
                "description": "Add a new supplier.",
                "operationId": "add_supplier_api_config_suppliers_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewSupplierRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupplierSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BOMItemSchema": {
                "properties": {
                    "material_name": {
                        "type": "string",
                        "title": "Material Name"
                    },
                    "quantity_required": {
                        "type": "number",
                        "title": "Quantity Required"
                    },
                    "pcb_ref": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Pcb Ref"
                    }
                },
                "type": "object",
                "required": [
                    "material_name",
                    "quantity_required"
                ],
                "title": "BOMItemSchema"
            },
...
            "ConfigResponse": {
                "properties": {
                    "models": {
                        "items": {
                            "$ref": "#/components/schemas/ProductModelSchema"
                        },
                        "type": "array",
                        "title": "Models"
                    },
                    "suppliers": {
                        "items": {
                            "$ref": "#/components/schemas/SupplierSchema"
                        },
                        "type": "array",
                        "title": "Suppliers"
                    }
                },
                "type": "object",
                "required": [
                    "models",
                    "suppliers"
                ],
                "title": "ConfigResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "NewSupplierRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "lead_time_days": {
                        "type": "integer",
                        "title": "Lead Time Days"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "lead_time_days"
                ],
                "title": "NewSupplierRequest"
            },
...
            "ProductModelSchema": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "assembly_time_days": {
                        "type": "integer",
                        "title": "Assembly Time Days"
                    },
                    "bom": {
                        "items": {
                            "$ref": "#/components/schemas/BOMItemSchema"
                        },
                        "type": "array",
                        "title": "Bom",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "assembly_time_days"
                ],
                "title": "ProductModelSchema"
            },
...
            "SupplierProductSchema": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "product_name": {
                        "type": "string",
                        "title": "Product Name"
                    },
                    "base_unit_cost": {
                        "type": "number",
                        "title": "Base Unit Cost"
                    },
                    "packaging_unit": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Packaging Unit"
                    },
                    "packaging_qty": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Packaging Qty"
                    },
                    "discount_tiers": {
                        "items": {},
                        "type": "array",
                        "title": "Discount Tiers",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "product_name",
                    "base_unit_cost"
                ],
                "title": "SupplierProductSchema"
            },
            "SupplierSchema": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "lead_time_days": {
                        "type": "integer",
                        "title": "Lead Time Days"
                    },
                    "active": {
                        "type": "boolean",
                        "title": "Active"
                    },
                    "products": {
                        "items": {
                            "$ref": "#/components/schemas/SupplierProductSchema"
                        },
                        "type": "array",
                        "title": "Products",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "lead_time_days",
                    "active"
                ],
                "title": "SupplierSchema"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "api/auth/login"
                    }
                }
            }
        }
    }
}